@Earnestly 2015-09-27
This does not seem to function with either: `./run_iwyu_tests.py -- include-what-you-use
` or `./run_iwyu_tests.py -- ./include-what-you-use`. Is the innvocation incorrect?