@kimgr 2015-09-27
Uh, oops. It searches the path before applying the command-line-provided path. I'll look into fixing that, but it will need a bigger effort, so it won't be today.