@Earnestly 2015-09-27
Yes, this works.  Thanks