@kimgr 2015-09-28
Here are some timings:

    Before
    ------
    Windows hog
        84 tests in 196.924s
        84 tests in 200.092s
        84 tests in 201.192s

    FreeBSD monster
        84 tests in 30.368s
        84 tests in 29.964s
        84 tests in 29.987s

    After
    -----
    Windows hog
        84 tests in 206.032s
        84 tests in 197.168s
        84 tests in 197.289s 

    FreeBSD monster
        84 tests in 31.948s
        84 tests in 30.404s
        84 tests in 30.014s

It looks like it might be slightly slower, but I'd say it's still within the noise.

I'll push the latest diff in a minute.