@Earnestly 2015-09-28
> “Any measurement that you make without the knowledge of its uncertainty is completely meaningless.”
— Walter Lewin