@vsapsai 2015-09-29
This change looks fine to me. It might be useful to tell about this option in Instructions for Developers