@kimgr 2015-09-29
Good idea, done.