@vsapsai 2015-09-30
Docs update is good, have nothing else to add.