@kimgr 2015-09-30
OK, I'll create a new branch and clean up commit history a little bit, and then push it. Thanks!